+++ doc/changelog.rst
@@ -12,6 +12,8 @@ Unrealeased
 Core
 ~~~~
 
+- Add `get_supplier` for shop product to define the business logic of getting supplier for order/basket line
+- Add shops to supplier to define which shops the supplier is available for
 - Introduce settings provider through new provide key `shuup_settings_provider`
 - Breaking change: Admin Section receives the request object and get methods converted to classmethod
 - API: allow user to remove and clear basket coupons